Closes #129: reorganize optim into directory structure
