+++ flax/optim/base.py
@@ -68,18 +68,15 @@ import abc
 from typing import Any
 import warnings
 
-from . import jax_utils
-from . import serialization
-from . import struct
-from . import traverse_util
+from .. import jax_utils
+from .. import serialization
+from .. import struct
+from .. import traverse_util
 
 import jax
-from jax import lax
 import jax.numpy as jnp
 
-from .nn import base
-
-import numpy as onp
+from ..nn import base
 
 
 @struct.dataclass
@@ -506,418 +503,3 @@ class ModelParamTraversal(traverse_util.Traversal):
       new_dict[key] = value
     new_params = traverse_util.unflatten_dict(new_dict)
     return inputs.replace(params=new_params)
-
-
-@struct.dataclass
-class _GradientDescentHyperParams:
-  learning_rate: onp.ndarray
-
-
-class GradientDescent(OptimizerDef):
-  """Gradient descent optimizer."""
-
-  def __init__(self, learning_rate=None):
-    """Constructor for the GradientDescent optimizer.
-
-    Args:
-      learning_rate: the step size used to update the parameters.
-    """
-    hyper_params = _GradientDescentHyperParams(learning_rate)
-    super().__init__(hyper_params)
-
-  def init_param_state(self, param):
-    return ()
-
-  def apply_param_gradient(self, step, hyper_params, param, state, grad):
-    del step
-    assert hyper_params.learning_rate is not None, 'no learning rate provided.'
-    new_param = param - hyper_params.learning_rate * grad
-    return new_param, state
-
-
-@struct.dataclass
-class _MomentumHyperParams:
-  learning_rate: onp.ndarray
-  beta: onp.ndarray
-  weight_decay: onp.ndarray
-  nesterov: bool
-
-
-@struct.dataclass
-class _MomentumParamState:
-  momentum: onp.ndarray
-
-
-class Momentum(OptimizerDef):
-  """Momentum optimizer."""
-
-  def __init__(self, learning_rate=None, beta=0.9, weight_decay=0,
-               nesterov=False):
-    """Constructor for the Momentum optimizer.
-
-    Args:
-      learning_rate: the step size used to update the parameters.
-      beta: the coefficient used for the moving average of the
-        gradient (default: 0.9).
-      weight_decay: weight decay coefficient to apply (default: 0).
-      nesterov: whether to use Nesterov momentum (default: False).
-    """
-
-    hyper_params = _MomentumHyperParams(
-        learning_rate, beta, weight_decay, nesterov)
-    super().__init__(hyper_params)
-
-  def init_param_state(self, param):
-    return _MomentumParamState(jnp.zeros_like(param))
-
-  def apply_param_gradient(self, step, hyper_params, param, state, grad):
-    del step
-    assert hyper_params.learning_rate is not None, 'no learning rate provided.'
-    if hyper_params.weight_decay != 0:
-      grad += hyper_params.weight_decay * param
-    momentum = state.momentum
-    new_momentum = hyper_params.beta * momentum + grad
-    if hyper_params.nesterov:
-      d_p = grad + hyper_params.beta * new_momentum
-    else:
-      d_p = new_momentum
-    new_param = param - hyper_params.learning_rate * d_p
-    new_state = _MomentumParamState(new_momentum)
-    return new_param, new_state
-
-
-@struct.dataclass
-class _AdamHyperParams:
-  learning_rate: onp.ndarray
-  beta1: onp.ndarray
-  beta2: onp.ndarray
-  eps: onp.ndarray
-  weight_decay: onp.ndarray
-
-
-@struct.dataclass
-class _AdamParamState:
-  grad_ema: onp.ndarray
-  grad_sq_ema: onp.ndarray
-
-
-class Adam(OptimizerDef):
-  """Adam optimizer."""
-
-  def __init__(self,
-               learning_rate=None,
-               beta1=0.9,
-               beta2=0.999,
-               eps=1e-8,
-               weight_decay=0.0):
-    """Constructor for the Adam optimizer.
-
-    Args:
-      learning_rate: the step size used to update the parameters.
-      beta1: the coefficient used for the moving average of the
-        gradient (default: 0.9).
-      beta2: the coefficient used for the moving average of the
-        gradient magnitude (default: 0.999).
-      eps: the term added to the gradient magnitude estimate for
-        numerical stability.
-      weight_decay: AdamW style weight decay rate
-        (relative to learning rate).
-    """
-    hyper_params = _AdamHyperParams(learning_rate, beta1, beta2, eps,
-                                    weight_decay)
-    super().__init__(hyper_params)
-
-  def init_param_state(self, param):
-    return _AdamParamState(jnp.zeros_like(param), jnp.zeros_like(param))
-
-  def apply_param_gradient(self, step, hyper_params, param, state, grad):
-    assert hyper_params.learning_rate is not None, 'no learning rate provided.'
-    beta1 = hyper_params.beta1
-    beta2 = hyper_params.beta2
-    weight_decay = hyper_params.weight_decay
-    grad_sq = lax.square(grad)
-    grad_ema = beta1 * state.grad_ema + (1. - beta1) * grad
-    grad_sq_ema = beta2 * state.grad_sq_ema + (1. - beta2) * grad_sq
-
-    # bias correction
-    t = step + 1.
-    grad_ema_corr = grad_ema / (1 - beta1 ** t)
-    grad_sq_ema_corr = grad_sq_ema / (1 - beta2 ** t)
-
-    denom = jnp.sqrt(grad_sq_ema_corr) + hyper_params.eps
-    new_param = param - hyper_params.learning_rate * grad_ema_corr / denom
-    if weight_decay != 0.0:
-      new_param -= hyper_params.learning_rate * weight_decay * param
-    new_state = _AdamParamState(grad_ema, grad_sq_ema)
-    return new_param, new_state
-
-
-@struct.dataclass
-class _LARSHyperParams:
-  learning_rate: onp.ndarray
-  beta: onp.ndarray
-  weight_decay: onp.ndarray
-  trust_coefficient: onp.ndarray
-  eps: onp.ndarray
-  nesterov: bool
-
-
-@struct.dataclass
-class _LARSParamState:
-  momentum: onp.ndarray
-
-
-class LARS(OptimizerDef):
-  """Layerwise adaptive rate scaling (LARS) optimizer.
-  
-  See https://arxiv.org/abs/1708.03888
-  """
-
-  def __init__(self, learning_rate=None, beta=0.9, weight_decay=0,
-               trust_coefficient=0.001, eps=0, nesterov=False):
-    """Constructor for the LARS optimizer.
-
-    Args:
-      learning_rate: the step size used to update the parameters.
-      beta: the coefficient used for the moving average of the
-        gradient (default: 0.9).
-      weight_decay: weight decay coefficient to apply
-      trust_coefficient: coefficient for trust ratio computation
-        (default: 0.001).
-      eps: epsilon used for trust ratio computation (default: no epsilon).
-      nesterov: whether to use Nesterov momentum (default: False).
-    """
-
-    hyper_params = _LARSHyperParams(
-        learning_rate, beta, weight_decay, trust_coefficient, eps, nesterov)
-    super().__init__(hyper_params)
-
-  def init_param_state(self, param):
-    return _LARSParamState(jnp.zeros_like(param))
-
-  def apply_param_gradient(self, step, hyper_params, param, state, grad):
-    del step
-    assert hyper_params.learning_rate is not None, 'no learning rate provided.'
-
-    param_norm = jnp.linalg.norm(param)
-    grad_norm = jnp.linalg.norm(grad)
-    trust_ratio = hyper_params.trust_coefficient * param_norm / (
-        grad_norm + hyper_params.weight_decay * param_norm + hyper_params.eps)
-    clipped_trust_ratio = jnp.where(
-        param_norm + grad_norm > 0., trust_ratio, 1.)
-    scaled_lr = hyper_params.learning_rate * clipped_trust_ratio
-    if hyper_params.weight_decay != 0:
-      grad += hyper_params.weight_decay * param
-
-    scaled_grad = scaled_lr * grad
-    momentum = state.momentum
-    new_momentum = hyper_params.beta * momentum + scaled_grad
-    if hyper_params.nesterov:
-      d_p = scaled_grad + hyper_params.beta * new_momentum
-    else:
-      d_p = new_momentum
-    new_param = param - d_p
-    new_state = _LARSParamState(new_momentum)
-    return new_param, new_state
-
-
-@struct.dataclass
-class _LAMBHyperParams:
-  learning_rate: onp.ndarray
-  beta1: onp.ndarray
-  beta2: onp.ndarray
-  weight_decay: onp.ndarray
-  eps: onp.ndarray
-
-
-@struct.dataclass
-class _LAMBParamState:
-  grad_ema: onp.ndarray
-  grad_sq_ema: onp.ndarray
-
-
-class LAMB(OptimizerDef):
-  """Layerwise adaptive moments for batch (LAMB) optimizer.
-  
-  See https://arxiv.org/abs/1904.00962
-  """
-
-  def __init__(self, learning_rate=None, beta1=0.9, beta2=0.999, weight_decay=0,
-               eps=1e-6):
-    """Constructor for the LAMB optimizer.
-
-    Args:
-      learning_rate: the step size used to update the parameters.
-      beta1: the coefficient used for the moving average of the gradient
-        (default: 0.9).
-      beta2: the coefficient used for the moving average of the squared gradient
-        (default: 0.999).
-      weight_decay: weight decay coefficient to apply
-      eps: epsilon used for Adam update computation (default: 1e-6).
-    """
-
-    hyper_params = _LAMBHyperParams(
-        learning_rate, beta1, beta2, weight_decay, eps)
-    super().__init__(hyper_params)
-
-  def init_param_state(self, param):
-    return _LAMBParamState(jnp.zeros_like(param), jnp.zeros_like(param))
-
-  def apply_param_gradient(self, step, hyper_params, param, state, grad):
-    beta1 = hyper_params.beta1
-    beta2 = hyper_params.beta2
-    weight_decay = hyper_params.weight_decay
-    learning_rate = hyper_params.learning_rate
-
-    grad_sq = lax.square(grad)
-    grad_ema = beta1 * state.grad_ema + (1. - beta1) * grad
-    grad_sq_ema = beta2 * state.grad_sq_ema + (1. - beta2) * grad_sq
-
-    t = step + 1.
-    grad_ema_corr = grad_ema / (1. - beta1 ** t)
-    grad_sq_ema_corr = grad_sq_ema / (1. - beta2 ** t)
-
-    update = grad_ema_corr / (jnp.sqrt(grad_sq_ema_corr) + hyper_params.eps)
-
-    if weight_decay != 0.0:
-      update += weight_decay * param
-
-    param_norm = jnp.linalg.norm(param)
-    update_norm = jnp.linalg.norm(update)
-    trust_ratio = jnp.where(
-        param_norm + update_norm > 0., param_norm / update_norm, 1.)
-
-    new_param = param - trust_ratio * learning_rate * update
-    new_state = _LAMBParamState(grad_ema, grad_sq_ema)
-    return new_param, new_state
-
-
-@struct.dataclass
-class _WeightNormHyperParams:
-  inner: Any
-  wn_decay: onp.ndarray
-  wn_eps: onp.ndarray
-
-
-@struct.dataclass
-class _WeightNormParamState:
-  direction_state: Any
-  scale_state: Any
-  mult: onp.ndarray
-
-
-class WeightNorm(OptimizerDef):
-  """Adds weight normalization to an optimizer def.
-
-  See https://arxiv.org/abs/1602.07868
-  """
-
-  def __init__(self, wrapped_optimizer, wn_decay=0, wn_eps=1e-8):
-    """Constructor for a WeightNorm optimizer.
-
-    Weight vectors are decomposed as w = g * v/||v||_2, for scalar
-    scale parameter g, and raw weight vector v. The original optimizer is then
-    applied to the (g,v) parameterization and the updated parameters are
-    transformed back to w-space, i.e.
-    w,state --> (g,v) --(original optimizer)--> (g',v') --> w',state'
-
-    We assume the output axis of any kernel matrix is the last one,
-    as per the Tensorflow convention.
-
-    Args:
-      wrapped_optimizer: another OptimizerDef
-      wn_decay: apply l2 decay to the unnoralized weight vector
-      wn_eps: additive constant for stability of
-        the normalization (default: 1e-8).
-    """
-    hps = _WeightNormHyperParams(
-        wrapped_optimizer.hyper_params, wn_decay, wn_eps)
-    super().__init__(hps)
-    self.wrapped_optimizer = wrapped_optimizer
-
-  def update_hyper_params(self, **hyper_param_overrides):
-    decay = hyper_param_overrides.pop('wn_decay', self.hyper_params.wn_decay)
-    eps = hyper_param_overrides.pop('wn_eps', self.hyper_params.wn_eps)
-    inner = self.wrapped_optimizer.update_hyper_params(
-        **hyper_param_overrides)
-    return self.hyper_params.replace(inner=inner, wn_decay=decay, wn_eps=eps)
-
-  def init_state(self, params):
-    leaves, treedef = jax.tree_flatten(params)
-    directions, scales = zip(*(self._split_param(p) for p in leaves))
-    directions = treedef.unflatten(directions)
-    scales = treedef.unflatten(scales)
-    wn_params = {'direction': directions, 'scale': scales}
-    state = self.wrapped_optimizer.init_state(wn_params)
-    direction_state = state.param_states['direction']
-    scale_state = state.param_states['scale']
-    param_states = jax.tree_multimap(
-        lambda _, *args: _WeightNormParamState(*args),
-        params, direction_state, scale_state, scales)
-    return state.replace(param_states=param_states)
-
-  def apply_gradient(self, hyper_params, params, state, grads):
-    p_leaves, treedef = jax.tree_flatten(params)
-    s_leaves = treedef.flatten_up_to(state.param_states)
-    g_leaves = treedef.flatten_up_to(grads)
-    split_grads = zip(*(self._split_grad(p, s, g, hyper_params.wn_decay)
-                        for p, s, g in zip(p_leaves, s_leaves, g_leaves)))
-    d_p, d_s, d_g, s_p, s_s, s_g = [
-        jax.tree_unflatten(treedef, x) for x in split_grads]
-    wn_params = {'direction': d_p, 'scale': s_p}
-    wn_state = {'direction': d_s, 'scale': s_s}
-    wn_grads = {'direction': d_g, 'scale': s_g}
-    new_wn_params, new_state = self.wrapped_optimizer.apply_gradient(
-        hyper_params.inner, wn_params,
-        state.replace(param_states=wn_state), wn_grads)
-
-    directions = treedef.flatten_up_to(new_wn_params['direction'])
-    scales = treedef.flatten_up_to(new_wn_params['scale'])
-    new_params, mults = zip(*(self._merge_param(d, s, hyper_params.wn_eps)
-                              for d, s in zip(directions, scales)))
-    new_params = jax.tree_unflatten(treedef, new_params)
-    mults = jax.tree_unflatten(treedef, mults)
-
-    direction_state = new_state.param_states['direction']
-    scale_state = new_state.param_states['scale']
-    param_states = jax.tree_multimap(
-        lambda _, *args: _WeightNormParamState(*args),
-        params, direction_state, scale_state, mults)
-    return new_params, new_state.replace(param_states=param_states)
-
-  def _split_param(self, param):
-    if param.size > param.shape[-1]:
-      scale = jnp.sqrt(jnp.square(param).sum(
-          tuple(range(param.ndim-1)), keepdims=True))
-      direction = param / scale
-      return direction, scale
-    else:
-      return param, ()
-
-  def _merge_param(self, direction, scale, eps):
-    if direction.size > direction.shape[-1]:
-      norm = jnp.sqrt(jnp.square(direction).sum(
-          tuple(range(direction.ndim - 1)), keepdims=True))
-      mult = scale / (eps + norm)
-      param = direction * mult
-      return param, mult
-    else:
-      return direction, ()
-
-  def _split_grad(self, param, state, grad, decay):
-    """Split the gradient for the direction and scale."""
-    if param.size > param.shape[-1]:
-      red_dims = tuple(range(param.ndim-1))
-      direction = param / state.mult
-      norm = jnp.sqrt(jnp.square(param).sum(red_dims, keepdims=True))
-      scale = norm * jnp.sign(state.mult)
-      scale_grad = jnp.sum(
-          grad * direction, axis=red_dims, keepdims=True)
-      direction_grad = state.mult * (grad - scale_grad * direction)
-      if decay is not 0:
-        direction_grad = direction_grad + decay * direction
-      direction_info = direction, state.direction_state, direction_grad
-      scale_info = scale, state.scale_state, scale_grad
-      return direction_info + scale_info
-    else:
-      return (param, state.direction_state, grad, (), (), ())
